fix issue: delete pets with schedules
Pets with schedules were not getting deleted due to a foreign key
violation occuring at the database layer.

When adding or deleting a schedule, one is now redirected back to the
view pet details page and not the pets page.

Unit tests were updated to check for cases where we are deleting pets
with schedules.